requirements.txt
fastapi[uvicorn]
fastapi-restful
uvicorn
sentry-sdk[fastapi]


typing_inspect
ccdexplorer-fundamentals
ccdexplorer_schema_parser
# https://stackoverflow.com/questions/78999040/mongodb-importerror-with-new-pymongo-release

grpcio
grpcio-tools
grpclib
betterproto==2.0.0b5

aiohttp
pydantic==2.9.0
prometheus-fastapi-instrumentator
python-dotenv
asgi-ratelimit
redis
fastapi-login
passlib
bcrypt==4.0.1
python-dotenv
humanize
pymongo==4.8.0
paho-mqtt
aiomqtt